ref(hybrid): Adapt silo scripts to getsentry
